Replaced with better unreal project and added possession/unpossession animation
I accidently started working on an older version of this project and made some progress, so I decided to just replace this current version with that one since it was further along. This explains the amount of files being committed. I also implemented a better animation for possessing and un-possessing a skeleton to make it smoother and more visually appealing.