Working base gameplay state machine
I have now got a working gameplay state machine for the basic states (Idle/Moving, Jumping/InAir, and Inactive). I have also fixed the bug where the player would not be able to input any actions after unpossessing a skeleton. This was due to my misunderstanding on how adding a mapping context works. I thought a mapping context had to be added for each pawn however I now know that it is added to the controller instead which is being used throughout, and when it is coded in a pawn, the pawn script is just a means to access the controller. Also gave the player the ability to still look around while in the inactive state (possessing the skeleton). Still need to code and create the other states (Interacting, Pushing, Floating, and Climbing).