Implementing modular interactable objects
Started on implementing the functionality and needed classes for modular interactable objects. My plan is to have a base interactable object class implementing the interactable interface. The interactable object will hold references to whatever mesh and animations are needed for this object which can be set in each blueprint scene instance. When the player interacts with this object, if it can still be interacted with, it will call its implemented interaction function and because the object will most likely be linked to another object such as a activatable gate, the gates activation function will also be called.