Implemented skeleton hover and activatable objects
While possessing a skeleton that implements the "Magic" interface, if the player is in the air they can use the Hover ability. This puts the player into the "Hover" state which makes their Z velocity 0 and disables their gravity, causing them to float in the direction they jumped. Will need to make some changes to allow the player to move while hovering. Also created a derived class of activatable object which when linked to an interactable object, calls its Activate function.