Created layout of first level, fixed camera rotation on possession/unpossession and created Minion derived classes Basic and Mage
I have created a basic layout of the first level using placeholder cubes as the different meshes. This level will introduce the movement, possession, and basic interaction mechanics. I finally fixed the issue with the players camera resetting to its default position every time you would possess or unpossess a pawn. During these events, I just store the Controller's rotation before possession and apply it back to the Controller once it had possessed/unpossessed a different pawn. I also created two new classess, Minion_Basic and Minion_Mage for their skeleton types respectively. This was needed as the base minion class was being used for the basic minion type and thus needed specific interfaces, but because the other minion classes were derived from this base, they would inherit those interfaces as well which was problematic for the warrior skeleton as it had to be Strong/Heavy and not Weak, which the basic minion type is.